fix: hide dependency on 'midi' library
When using this library from a TypeScript project, the following error
crops up:

```
node_modules/launchpad.js/dist/launchpads/base/BaseLaunchpad.d.ts:2:18 - error TS7016: Could not find a declaration file for module 'midi'.
  Try `npm i --save-dev @types/midi` if it exists or add a new declaration (.d.ts) file containing `declare module 'midi';`

import midi from 'midi';
                 ~~~~~~
```

With the current code structure, the dependency on the `midi` library is
leaking out, but since no types for it exist on NPM consumers can't even
install it.

Hide `midi` by not exporting the `utils` library any more (moving to a
`private` directory to signal it should not be exported), and by turning
the member MIDI device fields on `LaunchpadBase` to `private` (which
hides them from subclasses and therefore removes the need for retaining
the `import`).